Fix OSX no longer opening pages in browser
Bug:
g:livedown_browser = "firefox" is the default.
This translates to a node `exec firefox <url>` in  `livedown`.
On OSX `firefox` is not on the path.

Solution:
g:livedown_browser is unset by default.
The `--browser` flag is not passed.
The system launcher (ex. `open` on OSX), is used to open the url

Commentary:
`livedown` has the assumption that `--browser` is optional, when it's not
provided, it uses the system's launcher (ex. `open` on OSX). `vim-livedown`
should be the same way.